requirements.txt
# ============================================
# Projet Majeur IA — Pipeline OCR Tableau Blanc
# ESAIP · IR4-S8 · 2025–2026
# ============================================

# --- Core ---
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# --- ML / Deep Learning ---
torch>=2.0.0
torchvision>=0.15.0

# --- OCR Engines ---
python-doctr[torch]>=0.7.0
pytesseract>=0.3.10
transformers>=4.30.0       # TrOCR

# --- Data Augmentation ---
albumentations>=1.3.0

# --- Evaluation ---
jiwer>=3.0.0               # CER / WER
scikit-learn>=1.3.0

# --- Experiment Tracking ---
mlflow>=2.5.0

# --- PDF Generation ---
reportlab>=4.0.0

# --- Visualization ---
matplotlib>=3.7.0
seaborn>=0.12.0

# --- Utilities ---
tqdm>=4.65.0
pyyaml>=6.0
